+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## [0.0.5]
+
+* Properely Store and call callbacks queried previously and when received from server.
+* Add missing cmds permissionAdd, permissionRemove, permissionDetails and AdminExt methods for them.
+* Add missing params for some methods.
+
+## [0.0.4]
+
+* Null Safety
+* Add Individual ZeroNet Message Types and methods (both Callback model and Future model)
+
 ## [0.0.3]
 
 * Breaking Changes for ZeroNet.connect() method, changed params required.